Human_gesture_recognition/src/requirements.txt
Django==2.1
djangorestframework==3.8.2
joblib==0.14.0
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
